FIXED old leagues issue with teamIds
In some older leagues that had different owners each year, the team ids are not incremental. In a 12 person league, for example, the team ids were {1, 2, 4, 5, 6, 7, 9, 10, 11, 12, 13, 14}. This commit adjusts that issue when building the league. Other bugs may exist, but the self.adjustIds will be able to patch those issues as they arise.