Add support for private league authentication
If ESPN credentials (swid and espn_s2) are unknown, a browser will open in Chrome using selenium. The default chrome profile will load, which will hopefully log the user in automatically. If not, they will be prompted to log in. The swid and espn_s2 are then grabbed from the cookies automatically and the league is built as normal.