Update authorize.py
Selenium now opens the espn home screen and checks if the browser has logged the user in automatically. If not, the profile button is clicked, revealing the "Log In" button that the user must press and submit. The cookies available to selenium are refreshed every 5 seconds thereafter until the swid and espn_s2 cookies are available.

Now checks that the swid and espn_s2 have actually been identified by the browser and are not None values.