Check for league privacy first
If no swid / espn_s2 credentials are given, a league is build under the assumption that it can be accessed. If the league is not accessible, then `get_credentials()` is called and `buildLeague()` is attempted again.

`get_credentials()` is now called from 'buildLeague` rather than the `__init__`  method of 'League`