fix: Don't assume ordinal team_ids
If a league has N teams, it is not necessarily true that the team IDs are (0, 1, 2, 3, ... N-1). Sometimes (I think if a league previously had more than N teams and then deleted some), there could be a team_id that is > N. This change gets rid of this naiive idea that team IDs are ordered.