fix: Teams scoring 0 in outcomes
When passing in an outcome for a what_if scenario during simulations, 0 points were attributed to those teams. This has significant consequences on tiebreaking scenarios, diminishing both teams' odds. This is remedied by attributing to both teams their average ppg over the last 4 weeks. Though this is still not perfect, as in the final week of the season there will be absolutes that should not exist.