fix: Overwrite `league.current_week` with matchup period
When the `League` object is instantiated, the `current_week` attribute it set based on the current `scoringPeriod` of the league, which represents the actual week number. This can get out of sync with the desired matchup period for leagues with matchups that last longer than one week. Because this codebase uses week as a proxy for matchup period everywhere, it is best to assign `current_week` to its corresponding matchup period.