fix: Reset tables when week changes
When LeaguePage is first loading, the different tables display a loading spinner until the respective API calls have fetched data. However, when the `selectedWeek` is updated, the loading spinner was not updated. Instead, the Week headers and URL were changed, and then a few seconds later (when the new data was fetched), the tables would update. This fixes that so that the tables display loading spinners when `selectedWeek` is updated.