fix: Patch `ZeroDivisionError` for leagues with Week 1 Byes
If a league has bye weeks, those team's team.outcomes will be populated with a 'U' and not included in the denominator when calculating the win percentage. However, if a team had a bye in week 1, then their total games played will sum to 0 and raise a ZeroDivisionError: float division by zero. This patches the issue until it can be fixed in `espn-api`