fix: Add a `PseudoPlayer` class
Some teams may elect not to have any TEs on their roster (or other position) even though their league requires it. In these cases, an error message
```traceback-python
list index out of range
```
is thrown. This fix adds a `PseudoPlayer` class as a placeholder so that functions like `get_best_lineup()` do not break when indexing an empty list.